Add multiple configs + compiler/viewer output
Downstream can now select which render/backend to use via `.override`.
Currently, everything is supported _except_ the Skia renderer, mainly
due to difficulties with getting the bindings to generate against a
prebuilt version of the library (see skia-testing branch for how that
works atm).

Additionally adds the compiler and viewer as separate derivations.